feat: add soft deletes

+++ app/Http/Controllers/EnrollementController.php
@@ -9,6 +9,7 @@ use App\Models\GradeLevel;
 use App\Models\Student;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response;
 
 class EnrollementController extends Controller
@@ -25,58 +26,63 @@ class EnrollementController extends Controller
     }
 
     /**
+     * Display a listing of the resource remove soft.
+     */
+    public function softList()
+    {
+        //
+        $response = Gate::inspect('softList', Enrollement::class);
+
+        if (!$response->allowed()) {
+            return response()->json([
+                "message" => $response->message()
+            ], Response::HTTP_FORBIDDEN);
+        }
+
+        $deletedEnrollements = Enrollement::onlyTrashed()->get();
+        $deletedEnrollementsDTO = EnrollementDTO::fromCollection($deletedEnrollements);
+        return response()->json($deletedEnrollementsDTO, Response::HTTP_OK);
+    }
+
+    /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request, AcademicYear $academicYear, GradeLevel $gradeLevel)
     {
-        // pensar en que el estudainte ingrese por su codigo de estudiante
-        // al igual que el teacher
         //
-        // $validate = Validator::make($request->all(), [
-        //     'name' => 'required|string|max:64',
-        //     'first_name' => 'required|string|max:32',
-        //     'second_name' => 'required|string|max:32',
-        //     'phone' => 'nullable|string|max:32',
-        //     'birth_date' => 'nullable|date',
-        //     'address' => 'required|string|max:128',
-        //     'dni' => 'required|string|max:8|unique:users,dni,',
-        //     'email' => 'required|email|unique:users,email,',
-        // ]);
+        $response = Gate::inspect('store', Enrollement::class);
 
-        // if ($validate->fails()) {
-        //     return response()->json($validate->errors(), Response::HTTP_UNPROCESSABLE_ENTITY);
-        // }
-
-        // $i = Student::count();
-        // $code = 'ST' . (int)date('Y') * 10000 + $i;
-        // $student = Student::create([
-        //     'code_student' => $code,
-        //     'role' => User::ROLE_STUDENT
-        // ]);
+        if (!$response->allowed()) {
+            return response()->json([
+                "message" => $response->message()
+            ], Response::HTTP_FORBIDDEN);
+        }
 
         $i = Student::count();
 
         if ($i == 0) {
             $code = 'ST' . (int)date('Y') * 10000 + $i;
         } else {
-            $c = Student::latest("code_student")->first()->code_student;
+            $c = Student::latest("id")->first()->user->code;
             $i = (int)substr($c, 2) + 1;
             $code = 'ST' . $i;
         }
 
         $student = Student::create([
-            'code_student' => $code,
             'role' => User::ROLE_STUDENT
         ]);
 
         $new_datos = $request->validated_data;
+        $new_datos['code'] = $code;
         $new_datos['password'] = $code . $new_datos['dni'];
 
         $student->user()->create($new_datos);
 
         if (is_null($student->user)) {
             $student->delete();
-            return response()->json(["message" => "The registration for student $request->first_name $request->second_name $request->name has not been created successfully"], Response::HTTP_UNPROCESSABLE_ENTITY);
+            return response()->json([
+                "message" => "The registration for student $request->first_name $request->second_name $request->name has not been created successfully"
+            ], Response::HTTP_UNPROCESSABLE_ENTITY);
         }
 
         $enrollement = new Enrollement();
@@ -85,7 +91,9 @@ class EnrollementController extends Controller
         $enrollement->grade_level_id = $gradeLevel->id;
         $enrollement->save();
 
-        return response()->json(["message" => "The registration for student $request->first_name $request->second_name $request->name has been created successfully"], Response::HTTP_CREATED);
+        return response()->json([
+            "message" => "The registration for student $request->first_name $request->second_name $request->name has been created successfully"
+        ], Response::HTTP_CREATED);
     }
 
     /**
@@ -107,12 +115,78 @@ class EnrollementController extends Controller
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Remove soft the specified resource from storage.
      */
-    public function destroy(Enrollement $enrollement)
+    public function softDestroy(Enrollement $enrollement)
     {
         //
+        $response = Gate::inspect('softDestroy', Enrollement::class);
+
+        if (!$response->allowed()) {
+            return response()->json([
+                "message" => $response->message()
+            ], Response::HTTP_FORBIDDEN);
+        }
+
         $enrollement->delete();
-        return response()->json(["message" => "The enrollement has been deleted successfully"], Response::HTTP_ACCEPTED);
+        return response()->json([
+            "message" => "the enrollement with code $enrollement->id has been successfully deleted"
+        ], Response::HTTP_ACCEPTED);
+    }
+
+    /**
+     * Restore the specified resource from storage.
+     */
+    public function restore($id)
+    {
+        //
+        $response = Gate::inspect('restore', Enrollement::class);
+
+        if (!$response->allowed()) {
+            return response()->json([
+                "message" => $response->message()
+            ], Response::HTTP_FORBIDDEN);
+        }
+
+        $enrollement = Enrollement::onlyTrashed()->find($id);
+
+        if (is_null($enrollement)) {
+            return response()->json([
+                "message" => "the enrollement does not exist"
+            ], Response::HTTP_BAD_REQUEST);
+        }
+
+        $enrollement->restore();
+        return response()->json([
+            "message" => "the enrollement with code $enrollement->id has been successfully restored"
+        ], Response::HTTP_OK);
+    }
+
+    /**
+     * Remove permanently the specified resource from storage.
+     */
+    public function destroy($id)
+    {
+        //
+        $response = Gate::inspect('destroy', Enrollement::class);
+
+        if (!$response->allowed()) {
+            return response()->json([
+                "message" => $response->message()
+            ], Response::HTTP_FORBIDDEN);
+        }
+
+        $enrollement = Enrollement::onlyTrashed()->find($id);
+
+        if (is_null($enrollement)) {
+            return response()->json([
+                "message" => "the enrollement does not exist"
+            ], Response::HTTP_BAD_REQUEST);
+        }
+
+        $enrollement->forceDelete();
+        return response()->json([
+            "message" => "the enrollement with code $enrollement->id has been successfully deleted permanently"
+        ], Response::HTTP_ACCEPTED);
     }
 }
